Add files via upload
We have changed the H2 input from the OMVariable costs to be a percentage of the CAPEX (instead of a $/kg value). 
The LCOH calculation should also include (apart from CAPEX and OPEX): the costs of buying electricity from the grid. This is done via the spot price and the H2 electricity consumption.